Enum lightning_rapid_gossip_sync::GraphSyncError

pub enum GraphSyncError {
    DecodeError(DecodeError),
    LightningError(LightningError),
}

All-encompassing standard error type that processing can return

Variants

DecodeError(DecodeError)

Error trying to read the update data, typically due to an erroneous data length indication that is greater than the actual amount of data provided

LightningError(LightningError)

Error applying the patch to the network graph, usually the result of updates that are too old or missing prerequisite data to the application of updates out of order