Struct lightning_rapid_gossip_sync::RapidGossipSync

pub struct RapidGossipSync<NG: Deref<Target = NetworkGraph<L>>, L: Deref>where
    L::Target: Logger,{ /* private fields */ }

The main Rapid Gossip Sync object.

See crate-level documentation for usage.

Implementations

impl<NG: Deref<Target = NetworkGraph<L>>, L: Deref> RapidGossipSync<NG, L>where
L::Target: Logger,

pub fn new(network_graph: NG) -> Self

Instantiate a new RapidGossipSync instance.

pub fn sync_network_graph_with_file_path(
&self,
sync_path: &str
) -> Result<u32, GraphSyncError>

Sync gossip data from a file. Returns the last sync timestamp to be used the next time rapid sync data is queried.

network_graph: The network graph to apply the updates to

sync_path: Path to the file where the gossip update data is located

pub fn update_network_graph(
&self,
update_data: &[u8]
) -> Result<u32, GraphSyncError>

Update network graph from binary data. Returns the last sync timestamp to be used the next time rapid sync data is queried.

network_graph: network graph to be updated

update_data: &[u8] binary stream that comprises the update data

pub fn network_graph(&self) -> &NG

Gets a reference to the underlying NetworkGraph which was provided in RapidGossipSync::new.

(C-not exported) as bindings don’t support a reference-to-a-reference yet

pub fn is_initial_sync_complete(&self) -> bool

Returns whether a rapid gossip sync has completed at least once.